Only stderr print failed checks on other exception.
This tries to keep the overall output looking cleaner, and avoids
generating output that can make reading the results confusing.

It also resolves a corner case issue where a check is performed and
later retracted, but the printing of the failure had already occurred.

It's still possible for check stderr output to occur in the "wrong
spot" in the output of test results, but overall this should be
cleaner and clearer than the previous version.