Checking values does not introduce another withChecklist context.
The internal introduction of the internal context would cause an
evaluation of any check failures when that internal context was
exited, preventing the evaluation of the remainder of the test case.
This change removes that and updates the output information from
checks to note the input for _each_ failed check.